finished order endpoint

+++ src/main/java/com/cs322/ors/service/DishService.java
@@ -41,6 +41,9 @@ public class DishService {
 
 	public void deleteDish(long dishid) {
 		Optional<Dish> dish = this.dishRepository.findById(dishid);
-		dishRepository.delete(dish.get());
+		
+		if(dish.isPresent()){
+			dishRepository.delete(dish.get());
+		}
 	}
 }
